fix: split large useCharacterForm test file into smaller focused files
- Split 652-line useCharacterForm.test.ts into 3 files under 500-line limit:
  - useCharacterForm.basic.test.ts (165 lines) - Initial state and basic updates
  - useCharacterForm.classes-combat.test.ts (152 lines) - Classes and combat stats
  - useCharacterForm.validation.test.ts (349 lines) - Validation and edge cases
- Fix failing useCharacterSubmit test expectation for error details format
- All tests continue to pass (27 tests total)

Addresses Codacy file size limits for better code maintainability
